+++ Packages/com.frameworkbase.core/Tests/EditMode/SdkManagerTests.cs
@@ -126,6 +126,7 @@ namespace Framework.Tests
             public ISdkPrivacyService Privacy => null;
             public ISdkAdService Ad => null;
             public ISdkComplianceService Compliance => null;
+            public ISdkShareService Share => null;
 
             public UniTask<SdkResult> InitializeAsync() => UniTask.FromResult(SdkResult.Ok());